+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGELOG
 
+## 0.1.1
+
+* Use `sys_get_temp_dir()` and `tempnam()` to get a temporary file path if the
+  filesystem is local.
 ## 0.1.0
 
 * Initial Release